+++ src/forze/application/contracts/pubsub/__init__.py
@@ -1,4 +1,12 @@
+from .conformity import PubSubConformity, PubSubDepConformity
+from .deps import (
+    PubSubPublishDepKey,
+    PubSubPublishDepPort,
+    PubSubSubscribeDepKey,
+    PubSubSubscribeDepPort,
+)
 from .ports import PubSubPublishPort, PubSubSubscribePort
+from .specs import PubSubSpec
 from .types import PubSubMessage
 
 # ----------------------- #
@@ -7,4 +15,11 @@ __all__ = [
     "PubSubMessage",
     "PubSubPublishPort",
     "PubSubSubscribePort",
+    "PubSubConformity",
+    "PubSubDepConformity",
+    "PubSubSpec",
+    "PubSubPublishDepKey",
+    "PubSubPublishDepPort",
+    "PubSubSubscribeDepKey",
+    "PubSubSubscribeDepPort",
 ]